% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvData.R
\docType{data}
\name{cnvData}
\alias{cnvData}
\title{Pseudo Copy Number Variants Data}
\format{cnvData is a data.frame containing 522 observations with 5 columns: 
\describe{
\item{ID}{character patient identifier.}
\item{CHR}{CNV chromosome.}
\item{BP1}{starting location in base pairs.} 
\item{BP2}{ending location in base pairs.} 
\item{TYPE}{copy number (0,1,2,3,or 4).}
}}
\usage{
data(cnvData)
}
\description{
This data set includes simulated CNV data in PLINK CNV data format.
The data are also available from the authors through the url 
provided below. These data were generated following the simulation 
study used to illustrate the method in the original manuscript also 
referenced below; it has been reduced to include only 600 individuals. 
These data are not meaningful and are intended for demonstration purposes only.
}
\references{
Brucker, A., Lu, W., Marceau West, R., Yu, Q-Y., Hsiao, C. K.,
  Hsiao, T-H., Lin, C-H., Magnusson, P. K. E., Holloway, S. T., 
  Sullivan, P. F., Szatkiewicz, J. P., Lu, T-P., and
  Tzeng, J-Y. Association testing using Copy Number Profile Curves (CONCUR)
  enhances power in copy number variant analysis. <doi:10.1101/666875>.

\url{https://www4.stat.ncsu.edu/~jytzeng/Software/CONCUR/}
}
\keyword{datasets}
