% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvData.R
\docType{data}
\name{covData}
\alias{covData}
\title{Pseudo Covariate Data}
\format{covData is a data.frame containing 400 observations with 2 columns
\describe{
\item{ID}{character patient identifier.}
\item{SEX}{binary indicator of M/F.}
}}
\usage{
data(cnvData)
}
\description{
This data set includes simulated covariate data.
These data were generated as draws from a Binom(1,0.5) distribution for the
800 individuals in the example data provided with the package.
These data are not meaningful and are intended for demonstration purposes only.
}
\references{
Brucker, A., Lu, W., Marceau West, R., Yu, Q-Y., Hsiao, C. K.,
  Hsiao, T-H., Lin, C-H., Magnusson, P. K. E., Holloway, S. T., 
  Sullivan, P. F., Szatkiewicz, J. P., Lu, T-P., and
  Tzeng, J-Y. Association testing using Copy Number Profile Curves (CONCUR)
  enhances power in copy number variant analysis. <doi:10.1101/666875>.
}
\keyword{datasets}
