% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvData.R
\docType{data}
\name{phenoData}
\alias{phenoData}
\title{Pseudo Phenotype Data}
\format{
phenoData is a data.frame containing 400 observations with 3 columns
\describe{
\item{ID}{character patient identifier.}
\item{PHEB}{binary phenotype.}
\item{PHEC}{continuous phenotype.}
}
}
\usage{
data(cnvData)
}
\description{
This data set includes simulated phenotype data.
These data include a binary phenotype and a normally distributed continuous
phenotype that are randomly generated independent of the CNV data.
These data are not meaningful and are intended for demonstration purposes only.
}
\references{
Brucker, A., Lu, W., Marceau West, R., Yu, Q-Y., Hsiao, C. K.,
  Hsiao, T-H., Lin, C-H., Magnusson, P. K. E., Holloway, S. T., 
  Sullivan, P. F., Szatkiewicz, J. P., Lu, T-P., and
  Tzeng, J-Y. Association testing using Copy Number Profile Curves (CONCUR)
  enhances power in copy number variant analysis. <doi:10.1101/666875>.
}
\keyword{datasets}
